chore:refactor all utility functions into utilities code
