+++ tests/oot_test_config_models.py
@@ -14,11 +14,13 @@ import torch
 from pydantic import BaseModel, field_validator, model_validator  # type: ignore
 
 from oot_test_constants import (
+    _VALID_DTYPE_STRINGS,
+    _VALID_INIT_STRATEGIES,
+    _VALID_TEST_MODES,
+    _VALID_UNLISTED_MODES,
     DTYPE_STR_MAP,
     MODE_MANDATORY_SUCCESS,
-    MODE_SKIP,
     MODE_XFAIL,
-    MODE_XFAIL_STRICT,
     REL_PATH_TOKENS,
 )
 from oot_test_matching import parse_dtype
@@ -28,48 +30,6 @@ from oot_test_utilities import (
     _resolve_tensor_path,
 )
 
-
-# ---------------------------------------------------------------------------
-# Valid dtype strings (used in validators)
-# ---------------------------------------------------------------------------
-
-_VALID_DTYPE_STRINGS = {
-    "float16",
-    "float32",
-    "float64",
-    "bfloat16",
-    "int8",
-    "int16",
-    "int32",
-    "int64",
-    "uint8",
-    "uint16",
-    "uint32",
-    "uint64",
-    "complex32",
-    "complex64",
-    "complex128",
-    "bool",
-    "half",
-}
-# -------------------------------------------
-# Valid tensor generation strategies
-# -------------------------------------------
-_VALID_INIT_STRATEGIES = {
-    "rand",
-    "randn",
-    "zeros",
-    "ones",
-    "randint",
-    "arange",
-    "eye",
-    "full",
-    "file",
-}
-
-_VALID_TEST_MODES = {MODE_MANDATORY_SUCCESS, MODE_XFAIL, MODE_XFAIL_STRICT, MODE_SKIP}
-
-_VALID_UNLISTED_MODES = {"skip", "xfail", "xfail_strict", "mandatory_success"}
 
 # ---------------------------
 # edits.inputs models
